add archive manager
